# lib/api-zod/src/generated/types/adminCredentials.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AdminCredentials {
  /** @minLength 1 */
  password: string;
}