# lib/api-zod/src/generated/types/contactLead.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CreateContactLead } from './createContactLead';

export type ContactLead = CreateContactLead & {
  id: number;
  createdAt: Date;
};