# lib/api-zod/src/generated/types/createContactLead.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateContactLead {
  /**
     * @minLength 1
     * @maxLength 120
     */
  name: string;
  /** @maxLength 254 */
  email: string;
  /**
     * @minLength 1
     * @maxLength 5000
     */
  message: string;
}